Remove vault folder from wasm-internal
