+++ crates/bitwarden-vault/src/cipher/attachment_client.rs
@@ -10,6 +10,10 @@ use crate::{
     Cipher, DecryptError, EncryptError, VaultClient,
 };
 
+#[cfg(feature = "wasm")]
+use wasm_bindgen::prelude::*;
+
+#[cfg_attr(feature = "wasm", wasm_bindgen)]
 pub struct AttachmentsClient {
     pub(crate) client: Client,
 }
@@ -34,6 +38,24 @@ pub enum DecryptFileError {
     Io(#[from] std::io::Error),
 }
 
+#[cfg_attr(feature = "wasm", wasm_bindgen)]
+impl AttachmentsClient {
+    pub fn decrypt_buffer(
+        &self,
+        cipher: Cipher,
+        attachment: AttachmentView,
+        encrypted_buffer: &[u8],
+    ) -> Result<Vec<u8>, DecryptError> {
+        let key_store = self.client.internal.get_key_store();
+
+        Ok(key_store.decrypt(&AttachmentFile {
+            cipher,
+            attachment,
+            contents: EncString::from_buffer(encrypted_buffer)?,
+        })?)
+    }
+}
+
 impl AttachmentsClient {
     pub fn encrypt_buffer(
         &self,
@@ -49,6 +71,7 @@ impl AttachmentsClient {
             contents: buffer,
         })?)
     }
+
     pub fn encrypt_file(
         &self,
         cipher: Cipher,
@@ -65,20 +88,6 @@ impl AttachmentsClient {
         Ok(attachment)
     }
 
-    pub fn decrypt_buffer(
-        &self,
-        cipher: Cipher,
-        attachment: AttachmentView,
-        encrypted_buffer: &[u8],
-    ) -> Result<Vec<u8>, DecryptError> {
-        let key_store = self.client.internal.get_key_store();
-
-        Ok(key_store.decrypt(&AttachmentFile {
-            cipher,
-            attachment,
-            contents: EncString::from_buffer(encrypted_buffer)?,
-        })?)
-    }
     pub fn decrypt_file(
         &self,
         cipher: Cipher,
@@ -93,6 +102,7 @@ impl AttachmentsClient {
     }
 }
 
+#[cfg_attr(feature = "wasm", wasm_bindgen)]
 impl VaultClient {
     pub fn attachments(&self) -> AttachmentsClient {
         AttachmentsClient {
