+++ crates/bitwarden-wasm-internal/src/lib.rs
@@ -6,10 +6,8 @@ mod custom_types;
 mod init;
 mod pure_crypto;
 mod ssh;
-mod vault;
 
 pub use bitwarden_ipc::wasm::*;
 pub use client::BitwardenClient;
 pub use crypto::CryptoClient;
 pub use init::init_sdk;
-pub use vault::VaultClient;
